# swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Product API",
    "license": {
      "name": "Private"
    }
  },
  "host": "localhost:8081",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "definitions": {
    "User": {
      "properties": {
        "id": {
          "type": "string",
          "example": "1a36ade5-dfd4-46a9-a83e-9bbc803a2fb1",
          "description": "Unique record identifier"
        },
        "name": {
          "type": "string",
          "example": "John Doe",
          "description": "Agent's full-name"
        },
        "email": {
          "type": "string",
          "example": "john@company.com",
          "description": "Agent's e-mail"
        },
        "phoneNumber": {
          "type": "string",
          "example": 14086418460,
          "description": "Agent's phone-number (used for OTP verification)"
        },
        "avatar": {
          "type": "string",
          "example": "https://api.adorable.io/avatars/285/abott@adorable.png",
          "description": "Agent's avatar"
        },
        "roles": {
          "type": "array",
          "items": {
            "type": "string",
            "example": "sales"
          },
          "description": "Array of role name"
        },
        "createdAt": {
          "type": "string",
          "example": 1539626028333,
          "description": "Document's created timestamp"
        },
        "updatedAt": {
          "type": "string",
          "example": 1539626028333,
          "description": "Document's last updated timestamp"
        },
        "deletedAt": {
          "type": "string",
          "example": 1539626028333,
          "description": "Document's deleted timestamp"
        }
      }
    },
    "Product": {
      "properties": {
        "id": {
          "type": "string",
          "example": "1a36ade5-dfd4-46a9-a83e-9bbc803a2fb1",
          "description": "Unique record identifier"
        },
        "productName": {
          "type": "string",
          "example": "Bed Linen",
          "description": "Product full-name"
        },
        "productImage": {
          "type": "string",
          "example": "john@company.com",
          "description": "Product Image"
        },
        "productCategory": {
          "type": "string",
          "example": ".HOME FURNISHINGS",
          "description": "Product category Name"
        },
        "productPrice": {
          "type": "string",
          "example": "8000",
          "description": "Product Price"
        },
        "productDescription": {
          "type": "string",
          "example": "",
          "description": "Product Description"
        },
        "createdAt": {
          "type": "string",
          "example": 1539626028333,
          "description": "Document's created timestamp"
        },
        "updatedAt": {
          "type": "string",
          "example": 1539626028333,
          "description": "Document's last updated timestamp"
        },
        "deletedAt": {
          "type": "string",
          "example": 1539626028333,
          "description": "Document's deleted timestamp"
        }
      }
    }
  },
  "paths": {
    "/ping": {
      "tags": [
        "auth"
      ],
      "get": {
        "summary": "Check API",
        "operationId": "Check API",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/auth/register": {
      "post": {
        "summary": "Registration",
        "operationId": "PostRegisteration",
        "tags": [
          "auth"
        ],
        "parameters": [
          {
            "name": "data",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "required": [
                "email",
                "password",
                "name"
              ],
              "properties": {
                "email": {
                  "type": "string",
                  "description": "Login email"
                },
                "password": {
                  "type": "string",
                  "description": "Login password"
                },
                "name": {
                  "type": "string",
                  "description": "Name"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "summary": "Verify registration",
        "operationId": "PostLogin",
        "tags": [
          "auth"
        ],
        "parameters": [
          {
            "name": "data",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "required": [
                "email",
                "password"
              ],
              "properties": {
                "email": {
                  "type": "string",
                  "description": "Login email"
                },
                "password": {
                  "type": "string",
                  "description": "Login password"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/auth/update-password": {
      "post": {
        "summary": "update  password Agent",
        "operationId": "update password Agent",
        "tags": [
          "auth"
        ],
        "parameters": [
          {
            "name": "data",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "required": [
                "id",
                "password"
              ],
              "properties": {
                "id": {
                  "type": "string",
                  "description": "Agent's Id"
                },
                "password": {
                  "type": "string",
                  "description": "Agent new Password"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/customer"
            }
          }
        }
      }
    },
    "/product/add-product": {
      "post": {
        "summary": "Add Product",
        "operationId": "Post Add Product",
        "tags": [
          "product"
        ],
        "parameters": [
          {
            "name": "data",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "required": [
                "productName",
                "productCategory",
                "productPrice"
              ],
              "properties": {
                "productName": {
                  "type": "string",
                  "description": "Product Name"
                },
                "productCategory": {
                  "type": "string",
                  "description": "Product Main  Category"
                },
                "productPrice": {
                  "type": "string",
                  "description": "Product Price"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/product/get-list": {
    
      "post": {
        "summary": "List Product",
        "operationId": "List Product",
        "tags": [
          "product"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "type": "string",
            "example": "Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJFbXBsb3llcl9...",
            "required": false
          },
          {
            "name": "data",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "required": [
                "offset",
                "limit"
                            ],
              "properties": {
                "offset": {
                  "type": "number",
                  "description": "Total number of data want to fetch"
                },
                "limit": {
                  "type": "number",
                  "description": "Number of limit"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/product/{product_id}": {
      "get": {
        "summary": "Get Product By Id",
        "operationId": "Product",
        "tags": [
          "product"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "type": "string",
            "example": "Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJFbXBsb3llcl9...",
            "required": false
          },
          {
            "name": "product_id",
            "in": "path",
            "type": "string",
            "example": "60e5bc3264d6e8f220c664e9",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "ok"
          }
        }
      }
    },
    "/product/{productID}": {
      "delete": {
        "summary": "Delete Product By Id",
        "operationId": "DeleteProduct",
        "tags": [
          "product"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "type": "string",
            "example": "Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJFbXBsb3llcl9...",
            "required": false
          },
          {
            "name": "productID",
            "in": "path",
            "type": "string",
            "example": "60e5bc3264d6e8f220c664e9",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "ok"
          }
        }
      }
    },
    "/product/update-product": {
      "put": {
        "summary": "Update Product By Id",
        "operationId": "ProductUpdate",
        "tags": [
          "product"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "type": "string",
            "example": "Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJFbXBsb3llcl9...",
            "required": false
          },
          {
            "name": "data",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "required": [
                "id"
              ],
              "properties": {
                "id":
                {
                  "type": "string",
                  "description": "Product ID"
                },
                "productName": {
                  "type": "string",
                  "description": "Product Name"
                },
                "productImage": {
                  "type": "string",
                  "description": "product Image"
                },
                "productCategory":
                {
                   "type": "string",
                  "description": "Main Category name"
                },
                "productPrice":
                {
                  "type": "string",
                  "description": "Product Price"
                },
                "productDescription":
                {
                  "type": "string",
                  "description": "Product Description"
                },
                "hashtags":
                {
                  "type": "array",
                  "items":{
                      "type":"string",
                      "description": "hashtag name"
                  }
                }

              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "ok"
          }
        }
      }
    },
    "/product/search-product": {
      "post": {
        "summary": "Search Product By Tag",
        "operationId": "ProductByTag",
        "tags": [
          "product"
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "type": "string",
            "example": "Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJFbXBsb3llcl9...",
            "required": false
          },
          {
            "name": "data",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "required": [
                "hashtags"
              ],
              "properties": {
                "hashtags":
                {
                  "type": "array",
                  "items":{
                      "type":"string",
                      "description": "hashtag name"
                  }
                }

              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "ok"
          }
        }
      }
    },
    "/agent/profile": {
      "get": {
        "summary": "Get agent profile",
        "operationId": "GetProfile",
        "tags": [
          "agent"
        ],
        "parameters": [
          {
            "name": "email",
            "in": "query",
            "type": "string",
            "description": "Email of the agent"
          },
          {
            "name": "platform",
            "in": "query",
            "type": "string",
            "description": "platfrom of the agent"
          },
          {
            "name": "secretKey",
            "in": "query",
            "type": "string",
            "description": "secretKey of the agent"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Agent"
            }
          }
        }
      }
    }
  }
}